// backend/src/common/message.enum.ts
export enum Enum {
    NOT_FOUND = 'Registro no encontrado',    
    USER_NOT_FOUND = 'Usuario no encontrado',    
    INCORRECT_PASSWORD = 'Contraseña Incorrecta',    
    NOT_AUTHORIZED = 'No Autorizado',    
    USER_CONFLICT = 'Nombre de usuario no disponible',
    USER_CODE_CONFLICT = 'Código de usuario no disponible',    
}